requirements.txt
django==4.0.5
djangorestframework==3.13.1
django-cors-headers==3.12.0
psycopg2-binary==2.9.3
celery==5.2.7
redis==4.3.3
supervisor==4.2.4
gunicorn==20.1.0
boto3==1.24.51

pytest==7.1.2
pytest-django==4.5.2
pytest-cov==3.0.0
codecov==2.1.12

black==22.3.0
flake8==4.0.1
isort==5.10.1
